Fix consistency of color error test
rand() is not deterministic across platforms/implementations, so even
though the TestOneColourRandom() portion of the test was not
initializing the seed with a different value each time, the test result
would be different between MSVC and GCC for example.

It turns out that the test always has the same maximum error value
regardless of what the average is due to randomness. So instead, use
time() to reseed at the start of the test and then only check that the
max value is correct.