Don't try to find C component of OpenMP
None of the code in this library is pure C and as such no C compiler
is ever used; therefore, all of the 'C' specific aspects of CMake's
FindOpenMP module are irrelevant and trying to find them actually
causes OpenMP support to break.

This is because the original CMake scripts, which were for < 3.0
did not make use of the `LANGUAGES` part of the project() command,
as it did not exist. FindOpenMP automatically considers a language
component to not be found if that language was not specified to be
a part of the project via `project()` or `enable_language()`, both
of which did not exist prior to 3.0. Since we now explicitly enable
only the `CXX` language, OpenMP was always considered to be
"NOT_FOUND".

Drop checks for the 'C' component since it is not used anyway.